Extract ContentBlockHelper to behavior module for easier overriding
